Break Relay.Marathon.{App, Task} into own files
